pyproject.toml
[project]
name = "anonator"
version = "1.0.0"
description = "GPU-accelerated video face anonymization application"
requires-python = ">=3.9"
dependencies = [
    "opencv-python-headless>=4.8.0",
    "Pillow>=10.0.0",
    "tkinterdnd2>=0.3.0",
    "imageio>=2.25.0",
    "imageio-ffmpeg>=0.4.6",
    "numpy<2",
    "pyinstaller>=6.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]

build = [
    "pyinstaller>=5.13.0",
]

[build-system]
requires = ["setuptools>=67.6", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
